William li

+++ SyntaxAppProjectFolder/app/src/main/java/com/example/syntaxappproject/EventDetail.java
@@ -2,6 +2,16 @@ package com.example.syntaxappproject;
 
 import android.location.Criteria;
 
+/**
+ * Model class representing the detail of an event in the SyntaxEvents application.
+ * <p>
+ * Stores event id, organizer uid, event name, description, location, capacity,
+ * geoReq, event start and end date, start and end registration period,
+ * wait list count, lottery criteria, and poster. This class is used
+ * as a Firestore data model and is serialized/deserialized directly via
+ * {@code DocumentSnapshot.toObject(EventDetail.class)}.
+ * </p>
+ */
 public class EventDetail {
     // ─── Identity ─────────────────────────────────────────────────────────────
     private String eventId;
@@ -29,9 +39,28 @@ public class EventDetail {
     // ─── Media ────────────────────────────────────────────────────────────────
     private String poster;
 
-
+    /**
+     * Required no-argument constructor for Firestore deserialization.
+     */
     public EventDetail() {}
 
+    /**
+     * Constructs a fully initialized Profile.
+     *
+     * @param eventId                       the id of the event
+     * @param name                          the name of the event
+     * @param description                   the description of the event
+     * @param location                      the location of the event
+     * @param capacity                      the capacity of the event
+     * @param geoReq                        {@code true} if the event require geolocation
+     * @param startingEventDate             the start date of the event
+     * @param endingEventDate               the end date of the event
+     * @param startingRegistrationPeriod    the start date of registration
+     * @param endingRegistrationPeriod      the end date of registration
+     * @param waitlistCount                 the number of people in the wait list
+     * @param lotteryCriteria               the criteria of lottery
+     * @param poster                        the poster of the event
+     */
     public EventDetail(String eventId, String name, String description, String location, long capacity, boolean geoReq,
                  String startingEventDate, String endingEventDate,
                  String startingRegistrationPeriod, String endingRegistrationPeriod,
@@ -53,72 +82,210 @@ public class EventDetail {
 
 
     // Event ID
+    /**
+     * Returns the event id.
+     *
+     * @return the event id
+     */
     public String getEventId() { return eventId; }
+    /**
+     * Sets the event id.
+     *
+     * @param eventId the event id to assign
+     */
     public void setEventId(String eventId) { this.eventId = eventId; }
 
 
     // ─── Event Info ───
-
     // Name
+    /**
+     * Returns the name of the event.
+     *
+     * @return the event name
+     */
     public String getName() { return name; }
+    /**
+     * Sets the name of the event.
+     *
+     * @param name the event name to assign
+     */
     public void setName(String name) { this.name = name; }
-
+    /**
+     * Returns the uid of the organizer of the event.
+     *
+     * @return the uid of the organizer
+     */
     public String getOrganizerUid() { return organizerUid; }
+    /**
+     * Sets the uid of the organizer of the event.
+     *
+     * @param organizerUid the uid of the organizer to assign
+     */
     public void setOrganizerUid(String organizerUid) { this.organizerUid = organizerUid; }
 
     // Description
+    /**
+     * Returns the description of the event.
+     *
+     * @return the event description
+     */
     public String getDescription() { return description; }
+    /**
+     * Sets the description of the event.
+     *
+     * @param description the event description to assign
+     */
     public void setDescription(String description) { this.description = description; }
 
     // Location
+    /**
+     * Returns the location of the event.
+     *
+     * @return the event location
+     */
     public String getLocation() { return location; }
+    /**
+     * Sets the location of the event.
+     *
+     * @param location the event location to assign
+     */
     public void setLocation(String location) { this.location = location; }
 
     // Capacity
+    /**
+     * Returns the capacity of the event.
+     *
+     * @return the event capacity
+     */
     public long getCapacity() { return capacity; }
-    public void setCapacity(int capacity) { this.capacity = capacity; }
+    /**
+     * Sets the capacity of the event.
+     *
+     * @param capacity the event capacity to assign
+     */
+    public void setCapacity(long capacity) { this.capacity = capacity; }
 
     // Geolocation Requirement
+    /**
+     * Returns whether the event require geolocation
+     *
+     * @return {@code true} if the event require geolocation
+     */
     public boolean isGeoReq() { return geoReq; }
+    /**
+     * Sets the requirement of geolocation of the event.
+     *
+     * @param geoReq {@code true} to assign geolocation require
+     */
     public void setGeoReq(boolean geoReq) { this.geoReq = geoReq; }
 
 
     // ─── Event Dates ───
 
     // Starting Event Date
+    /**
+     * Returns the state date of the event.
+     *
+     * @return the event start date
+     */
     public String getStartingEventDate() { return startingEventDate; }
+    /**
+     * Sets the state date of the event.
+     *
+     * @param startingEventDate the event start date to assign
+     */
     public void setStartingEventDate(String startingEventDate) { this.startingEventDate = startingEventDate; }
 
     // Ending Event Date
+    /**
+     * Returns the end date of the event.
+     *
+     * @return the event end date
+     */
     public String getEndingEventDate() { return endingEventDate; }
+    /**
+     * Sets the end date of the event.
+     *
+     * @param endingEventDate the event end date to assign
+     */
     public void setEndingEventDate(String endingEventDate) { this.endingEventDate = endingEventDate; }
 
 
     // ─── Registration Period ───
 
     // Starting Registration Period
+    /**
+     * Returns the start date of the registration.
+     *
+     * @return the registration start date
+     */
     public String getStartingRegistrationPeriod() { return startingRegistrationPeriod; }
+    /**
+     * Sets the start date of the registration.
+     *
+     * @param startingRegistrationPeriod the registration start date to assign
+     */
     public void setStartingRegistrationPeriod(String startingRegistrationPeriod) { this.startingRegistrationPeriod = startingRegistrationPeriod; }
 
     // Ending Registration Period
+    /**
+     * Returns the end date of the registration.
+     *
+     * @return the registration end date
+     */
     public String getEndingRegistrationPeriod() { return endingRegistrationPeriod; }
+    /**
+     * Sets the end date of the registration.
+     *
+     * @param endingRegistrationPeriod the registration end date to assign
+     */
     public void setEndingRegistrationPeriod(String endingRegistrationPeriod) { this.endingRegistrationPeriod = endingRegistrationPeriod; }
 
 
     // ─── Lottery ───
 
     // Waitlist Count
+    /**
+     * Returns the count of people in the event wait list
+     *
+     * @return the count of people
+     */
     public long getWaitlistCount() { return waitlistCount; }
+    /**
+     * Sets the count of people in the event wait list
+     *
+     * @param waitlistCount the count of people to assign
+     */
     public void setWaitlistCount(long waitlistCount) { this.waitlistCount = waitlistCount; }
 
     // Lottery Criteria
+    /**
+     * Returns the criteria of lottery
+     *
+     * @return the lottery criteria
+     */
     public String getLotteryCriteria() { return lotteryCriteria; }
+    /**
+     * Sets the criteria of lottery
+     *
+     * @param lotteryCriteria the lottery criteria to assign
+     */
     public void setLotteryCriteria(String lotteryCriteria) { this.lotteryCriteria = lotteryCriteria; }
 
 
     // ─── Media ───
 
     // Poster
+    /**
+     * Returns the poster uri of the event
+     *
+     * @return the poster uri
+     */
     public String getPoster() { return poster; }
+    /**
+     * Sets the poster uri of the event
+     *
+     * @param poster the poster uri to assign
+     */
     public void setPoster(String poster) { this.poster = poster; }
 }
